requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
sqlalchemy==1.4.49
aiofiles==23.1.0
pdfminer.six
python-docx==0.8.11
sentence-transformers==2.2.2
faiss-cpu==1.12.0
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.21.0
